inputs are more lightweight, thanks to visitor pattern.
as usual, we can "use" inputs simply by suscribing them in a controller, but
now instead of using dynamic_cast(s), the Controller is kind of a Visitor.
Consequently, input types have to declare "bool handle(Controller * c)" with
"c->handle(this)", in order to be properly detected by the Controller.

Signed-off-by: Lasconik <caillouxnicolas@gmail.com>